[ADD] Selecting multiple ranges
